requirements.txt
google-api-python-client==2.95.0
grpcio-reflection==1.56.2
locust==2.29.0
pycountry==24.6.1
